requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pyngrok>=7.1.0
python-dotenv>=1.0.1
openai>=1.51.0
qdrant-client>=1.11.3
neo4j>=5.24.0
pydantic>=2.7.0
langchain-openai>=0.2.2
langchain>=0.3.7
typing-inspection>=0.4.0
